Implemented ValidationResult, to validate each model inside their own model

+++ StarCorp/Endpoints/CartEndpoints.cs
@@ -3,8 +3,10 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Routing;
 using StarCorp.Data;
+using StarCorp.Exceptions;
 using StarCorp.Models;
 using System;
+using System.ComponentModel.DataAnnotations;
 
 namespace StarCorp.Endpoints
 {
@@ -17,37 +19,55 @@ namespace StarCorp.Endpoints
 
             group.MapGet("/{cartId}", async (Guid cartId, ICartService cartService) =>
             {
-                var cart = await cartService.GetCartAsync(cartId);
+                if (cartId == Guid.Empty) return Results.BadRequest("Cart ID cannot be empty.");
 
-                if (cart == null)
+                try
                 {
-                    return Results.NotFound("Could not find the cart.");
+                    var cart = await cartService.GetCartAsync(cartId);
+                    return Results.Ok(cart);
                 }
-
-                return Results.Ok(cart);
+                catch (ResourceNotFoundException ex)
+                {
+                    return Results.NotFound(ex.Message);
+                }
             });
 
             group.MapPost("/{cartId}/items", async (Guid cartId, [FromBody] LineItem item, ICartService cartService) =>
             {
+                if (cartId == Guid.Empty) return Results.BadRequest("Cart ID cannot be empty.");
+
                 try
                 {
+                    ModelValidationException.ThrowIfInvalid(item);
+
                     await cartService.AddProductToCartAsync(cartId, item);
                     return Results.Ok();
                 }
-                catch (Exception ex)
+                catch (ValidationException ex)
                 {
                     return Results.BadRequest(ex.Message);
+                }
+                catch (ArgumentException ex)
+                {
+                    return Results.BadRequest(ex.Message);
+                }
+                catch (ResourceNotFoundException ex)
+                {
+                    return Results.NotFound(ex.Message);
                 }
             });
 
             group.MapDelete("/{cartId}/items/{productId}", async (Guid cartId, Guid productId, ICartService cartService) =>
             {
+                if (cartId == Guid.Empty) return Results.BadRequest("Cart ID cannot be empty.");
+                if (productId == Guid.Empty) return Results.BadRequest("Product ID cannot be empty.");
+
                 try
                 {
                     await cartService.RemoveProductFromCartAsync(cartId, productId);
                     return Results.Ok();
                 }
-                catch (Exception ex)
+                catch (ResourceNotFoundException ex)
                 {
                     return Results.NotFound(ex.Message);
                 }
